+++ wetterdienst/constants/access_credentials.py
@@ -1,5 +1,12 @@
 """ german weather service ftp credentials """
+from enum import Enum
+
+HTTPS_EXPRESSION = "https://"
+
 DWD_SERVER = "opendata.dwd.de"
 DWD_CDC_PATH = "climate_environment/CDC"
-DWD_CLIM_OBS_GERMANY_PATH = "observations_germany/climate"
-HTTPS_EXPRESSION = "https://"
+
+
+class DWDCDCBase(Enum):
+    CLIMATE_OBSERVATIONS = "observations_germany/climate"
+    GRIDS_GERMANY = "grids_germany"