+++ wetterdienst/enumerations/datetime_format_enumeration.py
@@ -7,3 +7,7 @@ class DatetimeFormat(Enum):
     YMDHM = "%Y%m%d%H%M"
     YMDH_COLUMN_M = "%Y%m%d%H:%M"
     YMD_TIME_H = "%Y-%m-%dT%H"
+
+    # For RADOLAN file datetime parsing
+    YM = "%Y%m"
+    ymdhm = "%y%m%d%H%M"