+++ wetterdienst/file_path_handling/file_list_creation.py
@@ -1,5 +1,6 @@
 """ file list creation for requested files """
 from typing import Union, List
+from datetime import datetime
 
 from wetterdienst.enumerations.column_names_enumeration import DWDMetaColumns
 from wetterdienst.enumerations.parameter_enumeration import Parameter
@@ -8,6 +9,7 @@ from wetterdienst.enumerations.time_resolution_enumeration import TimeResolution
 from wetterdienst.indexing.file_index_creation import (
     create_file_index_for_dwd_server,
     reset_file_index_cache,
+    create_file_index_for_radolan,
 )
 
 
@@ -48,3 +50,33 @@ def create_file_list_for_dwd_server(
     ]
 
     return file_index[DWDMetaColumns.FILENAME.value].values.tolist()
+
+
+def create_filepath_for_radolan(
+    date_time: datetime, time_resolution: TimeResolution
+) -> str:
+    """
+    Function used to create a relative filepath for a requested datetime depending on
+    the file index for the relevant time resolution.
+
+    Args:
+        date_time: datetime for requested RADOLAN file
+        time_resolution: time resolution enumeration of the request
+
+    Returns:
+        a string, either empty if non found or with the relative path to the file
+    """
+    file_index = create_file_index_for_radolan(time_resolution)
+
+    if date_time in file_index[DWDMetaColumns.DATETIME.value].tolist():
+        file_index = file_index[file_index[DWDMetaColumns.DATETIME.value] == date_time]
+    else:
+        file_index = file_index[
+            (file_index[DWDMetaColumns.DATETIME.value].dt.year == date_time.year)
+            & (file_index[DWDMetaColumns.DATETIME.value].dt.month == date_time.month)
+        ]
+
+    if file_index.empty:
+        return ""
+
+    return f"{file_index[DWDMetaColumns.FILENAME.value].item()}"