+++ wetterdienst/file_path_handling/path_handling.py
@@ -1,20 +1,25 @@
 """ functions to handle paths and file names"""
 from pathlib import Path, PurePosixPath
 from typing import Union, List
+from datetime import datetime
+
 from bs4 import BeautifulSoup
 
 from wetterdienst.constants.access_credentials import (
     HTTPS_EXPRESSION,
     DWD_SERVER,
     DWD_CDC_PATH,
-    DWD_CLIM_OBS_GERMANY_PATH,
+    DWDCDCBase,
 )
 from wetterdienst.constants.metadata import (
     DWD_FOLDER_STATION_DATA,
     DWD_FILE_STATION_DATA,
-    H5_FORMAT,
+    DataFormat,
+    DWD_FOLDER_RADOLAN,
+    DWD_FILE_RADOLAN,
 )
 from wetterdienst.download.https_handling import create_dwd_session
+from wetterdienst.enumerations.datetime_format_enumeration import DatetimeFormat
 from wetterdienst.enumerations.parameter_enumeration import Parameter
 from wetterdienst.enumerations.period_type_enumeration import PeriodType
 from wetterdienst.enumerations.time_resolution_enumeration import TimeResolution
@@ -46,8 +51,8 @@ def build_path_to_parameter(
     return parameter_path
 
 
-def list_files_of_climate_observations(
-    path: Union[PurePosixPath, str], recursive: bool
+def list_files_of_dwd(
+    path: Union[PurePosixPath, str], cdc_base: DWDCDCBase, recursive: bool
 ) -> List[str]:
     """
     A function used to create a listing of all files of a given path on the server
@@ -55,6 +60,7 @@ def list_files_of_climate_observations(
     Args:
         path: the path which should be searched for files (relative to climate
         observations Germany)
+        cdc_base: base path e.g. climate observations/germany
         recursive: definition if the function should iteratively list files
         from subfolders
 
@@ -63,7 +69,7 @@ def list_files_of_climate_observations(
     """
     dwd_session = create_dwd_session()
 
-    r = dwd_session.get(build_climate_observations_path(path))
+    r = dwd_session.get(build_dwd_cdc_data_path(path, cdc_base))
     r.raise_for_status()
 
     soup = BeautifulSoup(r.text, "html.parser")
@@ -83,7 +89,7 @@ def list_files_of_climate_observations(
 
     if recursive:
         files_in_folders = [
-            list_files_of_climate_observations(folder, recursive) for folder in folders
+            list_files_of_dwd(folder, cdc_base, recursive) for folder in folders
         ]
 
         for files_in_folder in files_in_folders:
@@ -92,22 +98,23 @@ def list_files_of_climate_observations(
     return files
 
 
-def build_climate_observations_path(path: Union[PurePosixPath, str]) -> str:
+def build_dwd_cdc_data_path(
+    path: Union[PurePosixPath, str], cdc_base: DWDCDCBase
+) -> str:
     """
     A function used to create the filepath consisting of the server, the
     climate observations path and the path of a subdirectory/file
 
     Args:
         path: the path of folder/file on the server
+        cdc_base: the CDC base path e.g. "climate observations/germany"
 
     Returns:
         the path create from the given parameters
     """
-    climate_observations_path = PurePosixPath(
-        DWD_SERVER, DWD_CDC_PATH, DWD_CLIM_OBS_GERMANY_PATH, path
-    )
+    dwd_cdc_data_path = PurePosixPath(DWD_SERVER, DWD_CDC_PATH, cdc_base.value, path)
 
-    return f"{HTTPS_EXPRESSION}{climate_observations_path}"
+    return f"{HTTPS_EXPRESSION}{dwd_cdc_data_path}"
 
 
 def build_local_filepath_for_station_data(folder: Union[str, Path]) -> Union[str, Path]:
@@ -121,7 +128,33 @@ def build_local_filepath_for_station_data(folder: Union[str, Path]) -> Union[str
         a Path build upon the folder
     """
     local_filepath = Path(
-        folder, DWD_FOLDER_STATION_DATA, f"{DWD_FILE_STATION_DATA}{H5_FORMAT}"
+        folder,
+        DWD_FOLDER_STATION_DATA,
+        f"{DWD_FILE_STATION_DATA}.{DataFormat.H5.value}",
+    ).absolute()
+
+    return local_filepath
+
+
+def build_local_filepath_for_radolan(
+    date_time: datetime, folder: Union[str, Path], time_resolution: TimeResolution
+) -> Union[str, Path]:
+    """
+
+    Args:
+        date_time:
+        folder:
+        time_resolution:
+
+    Returns:
+
+    """
+    local_filepath = Path(
+        folder,
+        DWD_FOLDER_RADOLAN,
+        time_resolution.value,
+        f"{DWD_FILE_RADOLAN}_{time_resolution.value}_"
+        f"{date_time.strftime(DatetimeFormat.YMDHM.value)}",
     ).absolute()
 
     return local_filepath